requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Pillow==10.0.0
reportlab==4.0.4
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Werkzeug==2.3.7
PyMuPDF==1.23.26
pytesseract==0.3.10
gunicorn==21.2.0 